voice-service/requirements.txt
# Speech Recognition
SpeechRecognition==3.13.0
PyAudio==0.2.14

# Text-to-Speech
pyttsx3==2.98

# Wake Word Detection
pvporcupine==3.0.6

# Audio Processing
numpy==2.2.1
scipy==1.15.1

# Utilities
python-dotenv==1.0.1
requests==2.32.3

# Logging
loguru==0.7.3